+++ cmd/restapi/apis/common.go
@@ -1,0 +1,5 @@
+package apis
+
+import "errors"
+
+var ErrEntityNotFound = errors.New("entity not found")
