+++ cmd/restapi/util/remote_db.go
@@ -1,4 +1,4 @@
-package main
+package util
 
 import (
 	"context"
